backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
mysql-connector-python>=8.2.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
google-auth>=2.23.0
pytest>=7.4.0
httpx>=0.25.0